+++ docs/releasenotes/releasenotes.rst
@@ -7,6 +7,7 @@ Latest changes
 Fixes
   * Improved error handling
   * Improved handling of chart data
+  * Almere garbage
 
 3.2.1 (10-12-2019)
 ------------------